email/milestone1/requirements.txt
langgraph
langchain
python-dotenv
